% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krwifi-package.R
\docType{package}
\name{krwifi-package}
\alias{krwifi}
\alias{krwifi-package}
\title{krwifi: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
