% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wifi.R
\docType{data}
\name{wifi}
\alias{wifi}
\title{wifi installation data in korea}
\format{A data frame with 21473 rows and 17 variables:
\describe{
\item{\code{설치장소명}}{character. DESCRIPTION.}
\item{\code{설치장소상세}}{character. DESCRIPTION.}
\item{\code{설치시도명}}{character. DESCRIPTION.}
\item{\code{설치시군구명}}{character. DESCRIPTION.}
\item{\code{설치시설구분}}{character. DESCRIPTION.}
\item{\code{서비스제공사명}}{character. DESCRIPTION.}
\item{\code{와이파이SSID}}{character. DESCRIPTION.}
\item{\code{설치년월}}{character. DESCRIPTION.}
\item{\code{소재지도로명주소}}{character. DESCRIPTION.}
\item{\code{소재지지번주소}}{character. DESCRIPTION.}
\item{\code{관리기관명}}{character. DESCRIPTION.}
\item{\code{관리기관전화번호}}{character. DESCRIPTION.}
\item{\code{위도}}{character. DESCRIPTION.}
\item{\code{경도}}{character. DESCRIPTION.}
\item{\code{데이터기준일자}}{character. DESCRIPTION.}
\item{\code{제공기관코드}}{character. DESCRIPTION.}
\item{\code{제공기관명}}{character. DESCRIPTION.}
}}
\usage{
wifi
}
\description{
from data.go.kr
}
\keyword{datasets}
